requirements.txt
celery==5.3.1
fastapi[standard]>=0.113.0,<0.114.0
python-dotenv==1.0.0
redis==5.1.1